Added New and Delete

+++ CosmosExplorer.Avalonia/ViewModels/MainWindowViewModel.cs
@@ -2,12 +2,15 @@
 using System;
 using System.Collections.ObjectModel;
 using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
 using CosmosExplorer.Core;
 using CosmosExplorer.Core.Models;
 using CosmosExplorer.Core.State;
 using DynamicData;
+using MsBox.Avalonia;
+using MsBox.Avalonia.Enums;
 
 namespace CosmosExplorer.Avalonia.ViewModels
 {
@@ -24,6 +27,7 @@ namespace CosmosExplorer.Avalonia.ViewModels
         private DocumentViewModel? selectedDocument;
         private DatabaseViewModel? selectedDatabase;
         private string? filter;
+        private string? query = "SELECT * FROM c";
         private PreferenceConnectionString selectedConnectionString;
         private string? connectionStringName;
         private string fullDocument;
@@ -39,6 +43,8 @@ namespace CosmosExplorer.Avalonia.ViewModels
 
         public ICommand GetDocumentAsyncCommand { get; }
         public ICommand ChangeConnectionStringCommand { get; }
+        
+        public ICommand DeleteConnectionStringAsyncCommand { get; }
         public ICommand ReLoadAsyncCommand { get; }
         
         public ObservableCollection<DocumentViewModel> Documents { get; set; } = [];
@@ -60,6 +66,7 @@ namespace CosmosExplorer.Avalonia.ViewModels
             DeleteAsyncCommand = ReactiveCommand.CreateFromTask(DeleteAsync); 
             GetDocumentAsyncCommand = ReactiveCommand.CreateFromTask(GetDocumentAsync);
             ChangeConnectionStringCommand = ReactiveCommand.CreateFromTask(ChangeConnectionStringAsync); 
+            DeleteConnectionStringAsyncCommand = ReactiveCommand.CreateFromTask(DeleteConnectionStringAsync); 
             ReLoadAsyncCommand = ReactiveCommand.CreateFromTask(ReloadAsync); 
             LoadSettingsAsyncCommand.Execute(null);
         }
@@ -96,12 +103,17 @@ namespace CosmosExplorer.Avalonia.ViewModels
             get => fullDocument;
             set => this.RaiseAndSetIfChanged(ref fullDocument, value);
         }
-        
-        public string? Query
+        public string? Filter
         {
             get => filter;
             set => this.RaiseAndSetIfChanged(ref filter, value);
         }
+        public string? Query
+        {
+            get => query;
+            set => this.RaiseAndSetIfChanged(ref query, value);
+        }
+        
         public DocumentViewModel? SelectedDocument
         {
             get => selectedDocument;
@@ -162,7 +174,6 @@ namespace CosmosExplorer.Avalonia.ViewModels
             SelectedConnectionString = loaded.ConnectionStrings.FirstOrDefault(c => c.Selected);
 
             ConnectionStrings = new (this.stateContainer.ConnectionStrings);
-            await ReloadAsync();
         }
 
         private async Task ReloadAsync()
@@ -184,6 +195,8 @@ namespace CosmosExplorer.Avalonia.ViewModels
                     }
                 }
                 if(Databases.Count>0) SelectedDatabase = Databases[0];
+                this.RaisePropertyChanged(nameof(SelectedDatabase)); 
+                
                 ReloadLastQueries();
             }
             catch (Exception e)
@@ -198,7 +211,7 @@ namespace CosmosExplorer.Avalonia.ViewModels
 
         private async Task SaveSettingsAsync()
         {
-            var addedConnectionString = new PreferenceConnectionString(connectionStringName, connectionString, true);
+            var addedConnectionString = new PreferenceConnectionString(connectionStringName, connectionString, true, false);
             stateContainer.ConnectionStrings.Add(addedConnectionString);
             await userSettingsService.SaveSettingsAsync(stateContainer);
             ConnectionStrings.Add(addedConnectionString);
@@ -208,7 +221,24 @@ namespace CosmosExplorer.Avalonia.ViewModels
             SelectedTabIndex = 0;
             AddConnectionString = false;
         }
-        
+
+        private async Task DeleteConnectionStringAsync()
+        {
+            var box = MessageBoxManager
+                .GetMessageBoxStandard("Delete?", "Are you sure you want to delete the connection string?",
+                    ButtonEnum.YesNo);
+
+            var result = await box.ShowAsync();
+            if (result == ButtonResult.Yes)
+            {
+                var index = stateContainer.ConnectionStrings.IndexOf(selectedConnectionString);
+                stateContainer.ConnectionStrings.RemoveAt(index);
+                await userSettingsService.SaveSettingsAsync(stateContainer);
+                ConnectionStringName = ConnectionString = "";
+                SelectedTabIndex = 0;
+                await LoadSettingsAsync();
+            }
+        }
         private async Task QueryAsync()
         {
             IsBusy = true;
@@ -224,7 +254,7 @@ namespace CosmosExplorer.Avalonia.ViewModels
                 Documents.AddRange(result.Select(x => new DocumentViewModel(x.Item1, x.Item2)));
                 AddLastQuery(filter);
                 await userSettingsService.SaveSettingsAsync(stateContainer);
-                Message = $"{count} items retrieved, {runits} RUs";
+                Message = $"{count} items retrieved, {runits:0.##} RUs";
 
             }
             catch (Exception e)
@@ -248,6 +278,48 @@ namespace CosmosExplorer.Avalonia.ViewModels
             try
             {
                 await SetContainerAsync();
+                Partition partition = cosmosDbDocumentService.Partition;
+                StringBuilder newDoc =
+                    new StringBuilder(
+                        """
+                        {
+                             "id": "replace_with_new_document_id",
+                        """);
+                if (partition is not null)
+                {
+                    newDoc.Append(
+                        $"""
+                         
+                              "{partition.PartitionName1}": "replace_with_new_partition_key_value"
+                         """
+                    );
+
+                    if (partition.PartitionName2 is not null)
+                    {
+                        newDoc.Append(
+                            $"""
+                             ,
+                                  "{partition.PartitionName2}": "replace_with_new_partition_key_value"
+                             """
+                        );
+                    }
+                    
+                    if (partition.PartitionName3 is not null)
+                    {
+                        newDoc.Append(
+                            $"""
+                             ,
+                                  "{partition.PartitionName3}": "replace_with_new_partition_key_value"
+                             """
+                        );
+                    }
+                    newDoc.Append("""
+                                  
+                                  }
+                                  """);
+                }
+
+                FullDocument = newDoc.ToString();
                 selectedDocument = null;
                 Documents.Clear();
             }
@@ -256,10 +328,11 @@ namespace CosmosExplorer.Avalonia.ViewModels
                 ErrorMessage = e.Message;
 
             }
+
             IsBusy = false;
 
         }
-        
+
         private async Task SaveAsync()
         {
             IsBusy = true;
@@ -268,33 +341,47 @@ namespace CosmosExplorer.Avalonia.ViewModels
             try
             {
                 await SetContainerAsync();
-                await cosmosDbDocumentService.UpdateDocumentAsync(selectedDocument.Id, selectedDocument.Partition, FullDocument);
+                
+                FullDocument = await cosmosDbDocumentService.UpdateDocumentAsync(selectedDocument?.Id, selectedDocument?.Partition, FullDocument);
             }
             catch (Exception e)
             {
                 ErrorMessage = e.Message;
 
             }
-            Message = "Document is updated.";
+                
+            Message = (selectedDocument is null)?"Document is added.":"Document is updated.";
             IsBusy = false;
 
         }
         private async Task DeleteAsync()
         {
-            IsBusy = true;
+            
             ErrorMessage = "";
             Message = "";
             try
             {
-                await SetContainerAsync();
-                await cosmosDbDocumentService.DeleteDocumentAsync(selectedDocument.Id, selectedDocument.Partition);
+                //await SetContainerAsync();
+                
+                var box = MessageBoxManager
+                    .GetMessageBoxStandard("Delete?", "Are you sure you want to delete the document?",
+                        ButtonEnum.YesNo);
+
+                var result = await box.ShowAsync();
+                if (result == ButtonResult.Yes)
+                {
+                    IsBusy = true;
+                    await cosmosDbDocumentService.DeleteDocumentAsync(selectedDocument.Id, selectedDocument.Partition);
+                    await QueryAsync();
+                    Message = "Document is deleted.";
+                }
             }
             catch (Exception e)
             {
                 ErrorMessage = e.Message;
 
             }
-            Message = "Document is deleted.";
+           
             IsBusy = false;
 
         }
